renamed files accordingly, removed action test file
